move processing logic from main
