+++ cmd/mdai-event-hub/main_test.go
@@ -3,14 +3,13 @@ package main
 import (
 	"os"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
 	dcorekube "github.com/decisiveai/mdai-data-core/kube"
+	"github.com/decisiveai/mdai-event-hub/internal/eventhub"
 	"github.com/decisiveai/mdai-event-hub/internal/handlers"
 	"github.com/decisiveai/mdai-event-hub/pkg/eventing"
-	operator "github.com/decisiveai/mdai-operator/api/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	valkeymock "github.com/valkey-io/valkey-go/mock"
@@ -79,7 +78,7 @@ func TestProcessEvent_Success(t *testing.T) {
 		"testHandler": testHandler,
 	})
 
-	handler := ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
+	handler := eventhub.ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
 	err = handler(event)
 
 	require.NoError(t, err)
@@ -130,7 +129,7 @@ func TestProcessEvent_NoHubName(t *testing.T) {
 		"testHandler": testHandler,
 	})
 
-	handler := ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
+	handler := eventhub.ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
 	err = handler(event)
 
 	require.Error(t, err)
@@ -182,7 +181,7 @@ func TestProcessEvent_MatchAlertNameOnly(t *testing.T) {
 		"testHandler": testHandler,
 	})
 
-	handler := ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
+	handler := eventhub.ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
 	err = handler(event)
 
 	require.NoError(t, err)
@@ -231,103 +230,11 @@ func TestProcessEvent_NoWorkflowFound(t *testing.T) {
 		"testHandler": testHandler,
 	})
 
-	handler := ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
+	handler := eventhub.ProcessAlertingEvent(ctx, mockClient, cmController, logger, nil, handlerMap)
 	err = handler(event)
 
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "no configured automation for event: UnknownAlert.firing")
-}
-
-func TestSafePerformAutomationStep_Success(t *testing.T) {
-	testHandler, state := newTestHandler()
-
-	handlerMap := handlers.GetSupportedHandlers(map[handlers.HandlerName]handlers.HandlerFunc{
-		"testHandler": testHandler,
-	})
-
-	mdai := handlers.MdaiInterface{
-		Logger: zap.NewNop(),
-	}
-
-	autoStep := operator.AutomationStep{
-		HandlerRef: "testHandler",
-		Arguments:  map[string]string{"key": "value"},
-	}
-
-	event := eventing.MdaiEvent{
-		HubName: "test-hub",
-		Name:    "TestAlert.firing",
-	}
-
-	err := safePerformAutomationStep(handlerMap, mdai, autoStep, event)
-
-	require.NoError(t, err)
-	assert.True(t, state.called)
-}
-
-func TestSafePerformAutomationStep_UnsupportedHandler(t *testing.T) {
-	mdai := handlers.MdaiInterface{
-		Logger: zap.NewNop(),
-	}
-
-	autoStep := operator.AutomationStep{
-		HandlerRef: "unsupportedHandler",
-		Arguments:  map[string]string{"key": "value"},
-	}
-
-	event := eventing.MdaiEvent{
-		HubName: "test-hub",
-		Name:    "TestAlert.firing",
-	}
-
-	testHandler, _ := newTestHandler()
-
-	handlerMap := handlers.GetSupportedHandlers(map[handlers.HandlerName]handlers.HandlerFunc{
-		"testHandler": testHandler,
-	})
-
-	err := safePerformAutomationStep(handlerMap, mdai, autoStep, event)
-
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "handler unsupportedHandler not supported")
-}
-
-func panicHandler(_ handlers.MdaiInterface, _ eventing.MdaiEvent, _ map[string]string) error {
-	panic("simulated panic")
-}
-
-func TestSafePerformAutomationStep_PanicRecovery(t *testing.T) {
-	handlerMap := handlers.GetSupportedHandlers(map[handlers.HandlerName]handlers.HandlerFunc{
-		"panicHandler": panicHandler,
-	})
-
-	logger := zap.NewNop()
-	mdai := handlers.MdaiInterface{
-		Data:   nil,
-		Logger: logger,
-	}
-
-	autoStep := operator.AutomationStep{
-		HandlerRef: "panicHandler",
-		Arguments:  map[string]string{},
-	}
-
-	event := eventing.MdaiEvent{
-		ID:      "1",
-		Name:    "TestEvent",
-		HubName: "test-hub",
-	}
-
-	err := safePerformAutomationStep(handlerMap, mdai, autoStep, event)
-	require.Error(t, err)
-
-	expectedPrefix := "panic in handler " + autoStep.HandlerRef
-	assert.True(
-		t,
-		strings.HasPrefix(err.Error(), expectedPrefix),
-		"unexpected error message:\n got: %q\n want prefix: %q",
-		err.Error(), expectedPrefix,
-	)
 }
 
 func TestProcessEventPayload_Success(t *testing.T) {
@@ -366,38 +273,6 @@ func TestProcessEventPayload_InvalidJSON(t *testing.T) {
 	assert.Nil(t, result, "expected result to be nil on invalid JSON")
 	require.Error(t, err, "expected an error for invalid JSON payload")
 	assert.Contains(t, err.Error(), "failed to unmarshal payload")
-}
-
-func TestGetWorkflowMap(t *testing.T) {
-	steps := []map[string]string{
-		{"TestAlert1.firing": "[{\"handlerRef\" :\"testHandler1\", \"args\": {\"key1\":\"value1\"}}]"},
-		{"TestAlert2.firing": "[{\"handlerRef\" :\"testHandler2\", \"args\": {\"key2\":\"value2\"}}]"},
-	}
-	expectedWorkflowMap := map[string][]operator.AutomationStep{
-		"TestAlert1.firing": {
-			{
-				HandlerRef: "testHandler1",
-				Arguments:  map[string]string{"key1": "value1"},
-			},
-		},
-		"TestAlert2.firing": {
-			{
-				HandlerRef: "testHandler2",
-				Arguments:  map[string]string{"key2": "value2"},
-			},
-		},
-	}
-	workflowMap := getRulesMap(zap.NewNop(), steps)
-	assert.Equal(t, expectedWorkflowMap, workflowMap)
-}
-
-func TestGetWorkflowMap_InvalidStepsJson(t *testing.T) {
-	steps := []map[string]string{
-		{"TestAlert1.firing": "[{\"hhhhandlerRef\" :\"testHandler1\", \"aaaargs\": {\"key1\":\"value1\"}}]"},
-	}
-	result := make(map[string][]operator.AutomationStep, 0)
-	workflowMap := getRulesMap(zap.NewNop(), steps)
-	assert.Equal(t, workflowMap, result)
 }
 
 func TestMain_CustomValkeyExpiryEnvVar(t *testing.T) {
